backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "VietNam transportation law RAG chatbot using FastAPI and LangChain"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    # --- Web Framework & Config ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",

    # --- LangChain Ecosystem (Theo phiên bản bạn cung cấp) ---
    "langchain~=0.3.25",
    "langchain-core~=0.3.65",
    "langchain-community~=0.3.25",
    "langchain-google-genai>=1.0.4",
    
    # --- AI Core & Data (THEO PHIÊN BẢN BẠN CUNG CẤP) ---
    "torch>=2.3.0", 
    "sentence-transformers~=4.1.0",
    "transformers",       # <-- Khớp chính xác
    "tokenizers",       # <-- Khớp chính xác
    "numpy>=2.1.0",
    "chromadb~=1.0.12",
    "pymupdf~=1.24.4",
    "rank-bm25>=0.2.2",
    "protobuf==3.20.3",

    # --- Database & Auth (for later stages) ---
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.1",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Khối tool.poetry để tương thích tốt nhất
[tool.poetry]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "VietNam transportation law RAG chatbot using FastAPI and LangChain"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]